Fix segfault when attempting to DeserializeArraySchema invalid data.
https://github.com/TileDB-Inc/TileDB/blob/ecfb3c629a8f3cf9377661afa94bedd0f26ec26f/tiledb/sm/c_api/tiledb.cc#L5275-L5290

In tiledb_serialize_array_schema, `*array_schema` will be
unconditionally set to a newly-allocated address. If an error later
occurs, `*array_schema` remains at that value, but the memory that it
points to *has already been freed*. That means that, on the Go side,
it sees the `ArraySchema.tiledbArraySchema` as a non-nil pointer,
so when it goes to free it in `ArraySchema.Free`, the memory will be
freed again.

This change works around that by only adding the finalizer *after*
it is verified that there was no error.