Rename tiledb-go command and do ASAN only when a tag is provided.
This renames the tiledb-go command to tiledb-go-examples, and makes it
build again by limiting ASAN to running when the user provides
a specific build tag (`-tags asan`).

In this case, running:

    go run ./cmd/tiledb-go-examples

will run the code without ASAN, but

    go run --tags=asan ./cmd/tiledb-go-examples

will run the code with ASAN by including `asan_on.go` rather than
`asan_off.go`.

https://pkg.go.dev/cmd/go#hdr-Build_constraints

Both `// +build` and `//go:build` comment-but-not-really-comments are
included in the code for forward compatibility.