Add test demonstrating Buffer's unsafety.
Buffer violates the CGo contract in two ways:

 1. It passes Go-managed memory into C, and C holds a pointer to that
    memory for longer than the duration of the call.
 2. When that memory *is* passed in, Buffer does not ensure that
    that memory remains alive. This means that it may be GC'd or
    overwritten before its next use.

Part 1 of the contract is currently not a huge concern: Go's GC does not
(yet?) actually relocate objects. In this case, so long as we hold up
Part 2, we should be fine. The problem is: we're not holding up Part 2,
and this test demonstrates that.