Ensure `bool`s are in every switch-case I can find. (#230)
I went through every (?) switch I could find in the whole library and
ensured that they supported `bool`s. In cases where it made sense,
I converted the switches to a shorter form and used generics to
dramatically reduce the amount of duplicated logic but with different
types.

Because we now use generics, we can't build on TileDB 1.17 (which is
fine, because our go.mod already specified that we need 1.18).

Also stops ignoring empty strings:

Previously, when a user tried to set an empty string in metadata,
the request would be ignored without any error. This (probably?) was
NOT the intended behavior.

Also removes unneeded nil-checks. Only a true nil (no type, no value)
would pass; a typed nil (yes type, no value) would not be caught.