Remove usage of deprecated tiledb methods to get query buffer sizes
In the latest release of TileDB the methods query_get_buffer(_nullable|_var)*
are deprecated and the recommended way to get the query buffers and their sizes is
to use a combination of get_data_buffer, get_offsets_buffer and get_validity_buffer

Note that for deserialized read queries, the result of get_(data|offsets|validity)_buffer
returns a nil pointer but a non-zero size. It is the callers responsibility to
first allocate buffers with the proper size and then run the query.