Return an error if the buffer is too large.
So far, we would cast a uint64 to a int32 but that may result in an
integer overflow if the uint64 cannot fit into the int32.  Since GoBytes
does not support buffer sizes of type uint64, the next best thing to do
is return an error if the given buffer is too large.